Ignore query options related to files
Using any option that causes the json results to contain keys with
file names causes problems for MongoDB. Currently the only ones are
ALL_FILES and CURRENT_FILES.

Mark those as unsupported and remove them from the options, if they
are given.

There is no other validation on the options, or combinations of
options, given in the configuration. If anything is invalid or not
supported by the Gerrit server, the REST call will fail.

Change-Id: If8c817dcc75872846316e2df3fdfc96236e495b5